requirements.txt
Flask==2.0.1
clickhouse-driver==0.2.9
pandas>=1.3.3
python-dotenv==0.19.0
pyjwt==2.1.0
Werkzeug==2.0.1 